requirements.txt
torch==2.1.2
torchvision==0.16.2
torchmetrics==1.3.0.post0
numpy==1.26.3
tqdm==4.66.1
pandas==2.2.0
matplotlib==3.8.2
matplotlib-inline==0.1.7
pillow==10.0.1
XAI-Library==0.0.4
ttach==0.0.3
lime==0.2.0.1
scikit-image==0.23.2
scikit-learn==1.4.0
scipy==1.12.0

